src/requirements.txt
alembic==0.8.6
amqp==1.4.9
anyjson==0.3.3
bcrypt==3.1.0
billiard==3.3.0.23
bleach==1.4.3
blinker==1.4
celery==3.1.23
cffi==1.7.0
click==6.6
cov-core==1.15.0
coverage==4.1
decorator==4.0.10
enum34==1.1.6
fake-factory==0.5.9
Flask==0.11.1
Flask-Admin==1.4.2
Flask-Bcrypt==0.7.1
Flask-Cache==0.13.1
Flask-Celery==2.4.3
Flask-Celery-Helper==1.1.0
Flask-Classy==0.6.10
Flask-DebugToolbar==0.10.0
flask-log==0.1.0
Flask-Migrate==1.8.0
Flask-Multi-Session==0.1.7
Flask-Script==2.0.5
Flask-Session==0.2.3
Flask-SQLAlchemy==2.1
gevent==1.1.1
greenlet==0.4.10
gunicorn==19.0.0
html5lib==0.9999999
infinity==1.4
intervals==0.7.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
Mako==1.0.4
MarkupSafe==0.23
MySQL-python==1.2.5
pathlib==1.0.1
pip-review==0.4
py==1.4.31
pycparser==2.14
PyMySQL==0.7.5
pytest==2.9.2
pytest-cov==2.3.0
pytest-flask==0.10.0
python-crfsuite==0.8.4
python-dateutil==2.5.3
python-editor==1.0.1
pytz==2016.4
PyYAML==3.11
redis==2.10.5
six==1.10.0
SQLAlchemy==1.0.14
SQLAlchemy-Utils==0.32.7
translitcodec==0.4.0
validators==0.10.3
webassets==0.11.1
Werkzeug==0.11.10
WTForms==2.1
